refactor(memory-audit): split into types/checks/auto-apply/report submodules
Brings packages/crane-mcp/src/tools/memory-audit.ts under structural caps
(was 663 LOC).

- memory-audit.ts (orchestrator, 175 LOC): thin entry that runs checks,
  builds summary, formats report.
- memory-audit/types.ts: interface/type definitions.
- memory-audit/checks.ts: the 7 audit check functions + flag builder.
- memory-audit/auto-apply.ts: promote-drafts + deprecate-zero-usage logic.
- memory-audit/report.ts: buildSummary + formatAuditReport + sub-formatters.

max-params (3 instances) + max-depth (1 instance) fixed via named-options
objects + extracted resolveSourcePath helper. Public API and re-exports of
MemoryAuditResult/MemoryAuditToolResult preserved.

10/10 tests pass.
